+++ typescript/ui/src/App.tsx
@@ -1,6 +1,6 @@
 import { Button, cn, ScrollShadow } from "@heroui/react";
 import Layout from "./core/components/Layout";
-import ChatMessage from "./core/components/ChatMessage";
+import { ChatMessage } from "./core/components/ChatMessage";
 import { useState, useRef, useEffect, useMemo, useCallback } from "react";
 import Markdown from "react-markdown";
 import remarkGfm from "remark-gfm";
@@ -8,6 +8,7 @@ import { Icon } from "@iconify/react";
 import { useConfigContext } from "./core/contexts/ConfigContext/useConfigContext";
 import { DEFAULT_LOGO, DEFAULT_SUBTITLE, DEFAULT_TITLE } from "./config";
 import {
+  useConversationProperty,
   useHistoryActions,
   useHistoryStore,
   useMessage,
@@ -20,10 +21,10 @@ export default function App() {
     config: { customization },
   } = useConfigContext();
   const messageIds = useMessageIds();
-  const lastMessageId = useHistoryStore((s) => s.lastMessageId);
+  const lastMessageId = useConversationProperty((s) => s.lastMessageId);
   const lastMessage = useMessage(lastMessageId);
   const historyIsLoading = useHistoryStore((s) => s.isLoading);
-  const followupMessages = useHistoryStore((s) => s.followupMessages);
+  const followupMessages = useConversationProperty((s) => s.followupMessages);
   const { sendMessage, stopAnswering } = useHistoryActions();
   const [showScrollDownButton, setShowScrollDownButton] = useState(false);
   const [shouldAutoScroll, setShouldAutoScroll] = useState(true);
@@ -107,13 +108,13 @@ export default function App() {
       <div className="flex w-full max-w-[600px] flex-col gap-4">
         {customization?.welcome_message && (
           <Markdown
-            className="text-center text-large text-default-900"
+            className="text-large text-default-900 text-center"
             remarkPlugins={[remarkGfm]}
           >
             {customization?.welcome_message}
           </Markdown>
         )}
-        <div className="text-center text-small text-default-500">
+        <div className="text-small text-default-500 text-center">
           You can start a conversation by typing in the input box below.
         </div>
       </div>
@@ -135,7 +136,7 @@ export default function App() {
   );
 
   return (
-    <div className="flex h-screen w-screen items-start justify-center bg-background">
+    <div className="bg-background flex h-screen w-screen items-start justify-center">
       <div className="h-full w-full max-w-full">
         <Layout subTitle={subTitle} title={title} logo={logo}>
           <div className="relative flex h-full flex-col overflow-y-auto p-6 pb-8">