+++ packages/ragbits-chat/CHANGELOG.md
@@ -1,11 +1,14 @@
 # CHANGELOG
 
 ## Unreleased
+- Update TailwindCSS, React, Vite, tailwind config (#742)
+- Add images support in chat message, images gallery (#731)
 - Add persistent user settings (#719)
 - Send chat options under `user_settings` key (#721)
 - Add feedback indicator to messages, allow `extensions` in chat messages (#722)
 - Added unit tests for UI's core components (#717)
 - Added share functionality with informative modal (#726)
+- Added persisent chat history to the default UI using IndexedDB (#732)
 
 ## 1.1.0 (2025-07-09)